Separate and refactor navigation
