+++ src/components/navigation/DrawerContent.js
@@ -1,0 +1,136 @@
+import React, { Component } from 'react'
+import {
+  ScrollView,
+  Image,
+  Text,
+  StyleSheet,
+  TouchableOpacity,
+  View
+} from 'react-native'
+import { withNamespaces } from 'react-i18next'
+import { connect } from 'react-redux'
+import PropTypes from 'prop-types'
+import { DrawerItems } from 'react-navigation'
+import CommunityIcon from 'react-native-vector-icons/MaterialCommunityIcons'
+import globalStyles from '../../globalStyles'
+import i18n from '../../i18n'
+import colors from '../../theme.json'
+import { switchLanguage } from '../../redux/actions'
+
+const mapStateToProps = ({ user }) => ({
+  user
+})
+
+const mapDispatchToProps = {
+  switchLanguage
+}
+
+// Component that renders the drawer menu content. DrawerItems are the links to
+// the given views.
+export class DrawerContent extends Component {
+  changeLanguage = lng => {
+    i18n.changeLanguage(lng)
+    this.props.switchLanguage(lng)
+    this.props.navigation.toggleDrawer()
+  }
+  render() {
+    const { lng, user } = this.props
+    return (
+      <ScrollView contentContainerStyle={styles.container}>
+        <View>
+          <Image
+            style={{ height: 172, width: 304 }}
+            source={require('../../../assets/images/navigation_image.png')}
+          />
+          <View style={styles.languageSwitch}>
+            <TouchableOpacity id="en" onPress={() => this.changeLanguage('en')}>
+              <Text
+                style={[
+                  globalStyles.h3,
+                  lng === 'en' ? styles.whiteText : styles.greyText
+                ]}
+              >
+                EN
+              </Text>
+            </TouchableOpacity>
+            <Text style={[globalStyles.h3, styles.whiteText]}>
+              {'  '}|{'  '}
+            </Text>
+            <TouchableOpacity id="es" onPress={() => this.changeLanguage('es')}>
+              <Text
+                style={[
+                  globalStyles.h3,
+                  lng === 'es' ? styles.whiteText : styles.greyText
+                ]}
+              >
+                ESP
+              </Text>
+            </TouchableOpacity>
+          </View>
+          <Text
+            id="username"
+            style={[styles.username, globalStyles.h3, styles.whiteText]}
+          >
+            {user.username}
+          </Text>
+          <DrawerItems {...this.props} />
+        </View>
+        <TouchableOpacity id="logout" style={styles.logout} onPress={() => {}}>
+          <CommunityIcon name="logout" size={20} color={colors.palegreen} />
+          <Text style={styles.logoutLabel}>Logout</Text>
+        </TouchableOpacity>
+      </ScrollView>
+    )
+  }
+}
+
+DrawerContent.propTypes = {
+  lng: PropTypes.string,
+  switchLanguage: PropTypes.func.isRequired,
+  navigation: PropTypes.object.isRequired,
+  user: PropTypes.object.isRequired
+}
+
+export default withNamespaces()(
+  connect(
+    mapStateToProps,
+    mapDispatchToProps
+  )(DrawerContent)
+)
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    flexDirection: 'column',
+    justifyContent: 'space-between'
+  },
+  greyText: {
+    color: colors.palegrey
+  },
+  whiteText: {
+    color: colors.white
+  },
+  languageSwitch: {
+    flexDirection: 'row',
+    position: 'absolute',
+    top: 40,
+    left: 16
+  },
+  username: {
+    position: 'absolute',
+    top: 139,
+    left: 16
+  },
+  logout: {
+    flexDirection: 'row',
+    marginLeft: 15,
+    marginBottom: 25
+  },
+  logoutLabel: {
+    marginLeft: 20,
+    fontFamily: 'Poppins',
+    fontWeight: '600',
+    fontSize: 14,
+    color: colors.palegreen
+  }
+})
