+++ src/components/navigation/LifemapStack.js
@@ -1,0 +1,15 @@
+import { createStackNavigator } from 'react-navigation'
+import SurveysView from '../../screens/Surveys'
+import LifemapScreens from './LifemapScreens'
+import { generateNavOptions } from './helpers'
+
+export default createStackNavigator({
+  Surveys: {
+    screen: SurveysView,
+    navigationOptions: ({ navigation }) => ({
+      title: 'Create a Life Map',
+      ...generateNavOptions({ navigation })
+    })
+  },
+  ...LifemapScreens
+})
