+++ src/components/navigation/SyncStack.js
@@ -1,0 +1,17 @@
+import { createStackNavigator } from 'react-navigation'
+import SyncView from '../../screens/Sync'
+import { generateNavOptions } from './helpers'
+
+export default createStackNavigator(
+  {
+    Sync: {
+      screen: SyncView,
+      navigationOptions: {
+        title: 'Sync'
+      }
+    }
+  },
+  {
+    navigationOptions: ({ navigation }) => generateNavOptions({ navigation })
+  }
+)
