+++ src/fake-socio-economic-data.json
@@ -1,0 +1,84 @@
+{
+  "surveyEconomicQuestions": [
+    {
+      "questionText": "Is there any member with disabilities in your household? Please indicate the disability type",
+      "answerType": "select",
+      "dimension": "Family details",
+      "required": true,
+      "forFamilyMember": false,
+      "options": [
+        { "value": "PHYSICAL", "text": "Phisical" },
+        { "value": "MENTAL", "text": "Mental" },
+        { "value": "LEARNING", "text": "Learning" },
+        {
+          "value": "NO-MEMBER-DISABILITIES",
+          "text": "No member with disabilities"
+        }
+      ]
+    },
+    {
+      "questionText": "What is the property title situation of your household?",
+      "answerType": "select",
+      "dimension": "Family details",
+      "required": false,
+      "forFamilyMember": false,
+      "options": [
+        { "value": "OWNER", "text": "Owner" },
+        {
+          "value": "COUNCIL-HOUSING-ASSOCIATION",
+          "text": "Council/Housing Association"
+        },
+        { "value": "PRIVATE-RENTAL", "text": "Private rental" },
+        { "value": "LIVING-WITH-PARENTS", "text": "Living with Parents" },
+        { "value": "PREFER-NOT-TO-SAY", "text": "Prefer not to say" }
+      ]
+    },
+    {
+      "questionText": "What is your highest educational level?",
+      "answerType": "select",
+      "dimension": "Education",
+      "required": true,
+      "forFamilyMember": true,
+      "options": [
+        { "value": "SCHOOL-INCOMPLETE", "text": "School incomplete" },
+        { "value": "SCHOOL-COMPLETE", "text": "Apprenticeship complete" },
+        { "value": "APPRENTICESHIP-COMPLETE", "text": "Unmployed" },
+        { "value": "COLLEGE-COMPLETE", "text": "College complete" },
+        { "value": "UNIVERSITY-COMPLETE", "text": "University complete" }
+      ]
+    },
+    {
+      "questionText": "What is your primary economic activity?",
+      "answerType": "select",
+      "dimension": "Income",
+      "required": true,
+      "forFamilyMember": false,
+      "options": [
+        { "value": "EMPLOYED", "text": "Employed" },
+        { "value": "SELF-EMPLOYED", "text": "Self Employed" },
+        { "value": "UNEMPLOYED", "text": "Unmployed" },
+        { "value": "RETIRED", "text": "Retired" },
+        { "value": "NONE", "text": "None" }
+      ]
+    },
+    {
+      "questionText": "Please estimate your gross monthly household income (i.e, before taxes National Insurance contributions or other deductions)",
+      "answerType": "text",
+      "dimension": "Income",
+      "required": true,
+      "forFamilyMember": false,
+      "options": []
+    },
+    {
+      "questionText": "Do you receive any state income. (e.g. allowances, benefits, disability, pension)?",
+      "answerType": "select",
+      "dimension": "Income",
+      "required": false,
+      "forFamilyMember": false,
+      "options": [
+        { "value": "YES", "text": "yes" },
+        { "value": "NO", "text": "no" }
+      ]
+    }
+  ]
+}
